requirements.txt
python-telegram-bot==20.7
langchain
google-generativeai
faiss-cpu
unstructured
tqdm
python-dotenv
PyPDF2
